@@ -95,20 +95,20 @@
     "request-promise": "4.2.4",
     "request-promise-native": "1.0.7",
     "scanf": "^1.0.2",
-    "sequelize": "4.42.1",
-    "sequelize-typescript": "0.6.7",
+    "sequelize": "^5.2.12",
+    "sequelize-typescript": "0.6.9",
     "shx": "0.3.2",
-    "simple-git": "1.107.0",
+    "simple-git": "1.110.0",
     "sqlite3": "4.0.6",
     "strict-password-generator": "^1.1.2",
     "swagger-client": "3.8.25",
-    "tedious": "5.0.3",
+    "tedious": "6.1.0",
     "temperature-js": "^0.1.0",
-    "ts-node": "8.0.2",
+    "ts-node": "8.0.3",
     "typedoc": "0.14.2",
     "typedoc-plugin-external-module-name": "^2.0.0",
     "typedoc-plugin-markdown": "^1.1.27",
-    "typescript": "3.3.3333",
+    "typescript": "3.4.1",
     "url-join": "4.0.0",
     "vbscript-to-typescript": "^1.0.8",
     "wait-until": "0.0.2",
