@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -37,39 +37,40 @@
 'use strict';
 
 import { CognitiveServicesManagementClient } from 'azure-arm-cognitiveservices';
-import { CognitiveServicesAccount } from 'azure-arm-cognitiveservices/lib/models';
 import { ResourceManagementClient, SubscriptionClient } from 'azure-arm-resource';
 import { SearchManagementClient } from 'azure-arm-search';
 import { SqlManagementClient } from 'azure-arm-sql';
 import { WebSiteManagementClient } from 'azure-arm-website';
+//tslint:disable-next-line:no-submodule-imports
 import { AppServicePlan } from 'azure-arm-website/lib/models';
-import { GBService, IGBInstance } from 'botlib';
-import { HttpMethods, ServiceClient, WebResource } from 'ms-rest-js';
-import { GBDeployer } from '../../../packages/core.gbapp/services/GBDeployer';
-import * as simplegit from 'simple-git/promise';
+import { GBLog, IGBInstallationDeployer, IGBInstance } from 'botlib';
+import { HttpHeaders, HttpMethods, ServiceClient, WebResource } from 'ms-rest-js';
 import { GBAdminService } from '../../../packages/admin.gbapp/services/GBAdminService';
 import { GBCorePackage } from '../../../packages/core.gbapp';
 import { GBConfigService } from '../../../packages/core.gbapp/services/GBConfigService';
-import { GuaribasInstance } from '../../../packages/core.gbapp/models/GBModel';
+import { GBDeployer } from '../../../packages/core.gbapp/services/GBDeployer';
 
 const Spinner = require('cli-spinner').Spinner;
-const scanf = require('scanf');
-const git = simplegit();
-const logger = require('../../../src/logger');
-const UrlJoin = require('url-join');
+
+// tslint:disable-next-line:no-submodule-imports
+import { CognitiveServicesAccount } from 'azure-arm-cognitiveservices/lib/models';
+import urlJoin = require('url-join');
 const iconUrl = 'https://github.com/pragmatismo-io/BotServer/blob/master/docs/images/generalbots-logo-squared.png';
 const publicIp = require('public-ip');
 
-export class AzureDeployerService extends GBService {
-  public static apiVersion = '2017-12-01';
-  public static defaultEndPoint = 'http://localhost:4242';
+/**
+ * Deployer for Microsoft cloud.
+ */
+export class AzureDeployerService implements IGBInstallationDeployer {
+  public apiVersion = '2017-12-01';
+  public defaultEndPoint = 'http://localhost:4242';
   public instance: IGBInstance;
   public resourceClient: ResourceManagementClient.ResourceManagementClient;
   public webSiteClient: WebSiteManagementClient;
   public storageClient: SqlManagementClient;
   public cognitiveClient: CognitiveServicesManagementClient;
   public searchClient: SearchManagementClient;
-  public static provider = 'Microsoft.BotService';
+  public provider = 'Microsoft.BotService';
   public subscriptionClient: SubscriptionClient.SubscriptionClient;
   public accessToken: string;
   public location: string;
@@ -78,15 +79,28 @@ export class AzureDeployerService extends GBService {
   public deployer: GBDeployer;
 
   constructor(deployer: GBDeployer) {
-    super();
     this.deployer = deployer;
   }
-  public static async getSubscriptions(credentials) {
+
+  private static createRequestObject(url: string, accessToken: string, verb: HttpMethods, body: string) {
+    const req = new WebResource();
+    req.method = verb;
+    req.url = url;
+    req.headers.set('Content-Type', 'application/json');
+    req.headers.set('accept-language', '*');
+    req.headers.set('Authorization', `Bearer ${accessToken}`);
+    req.body = body;
+
+    return req;
+  }
+
+  public async getSubscriptions(credentials) {
     const subscriptionClient = new SubscriptionClient.default(credentials);
+
     return subscriptionClient.subscriptions.list();
   }
 
-  public static getKBSearchSchema(indexName) {
+  public getKBSearchSchema(indexName) {
     return {
       name: indexName,
       fields: [
@@ -182,21 +196,18 @@ export class AzureDeployerService extends GBService {
         }
       ],
       scoringProfiles: [],
-      defaultScoringProfile: null,
-      corsOptions: null
+      defaultScoringProfile: undefined,
+      corsOptions: undefined
     };
   }
 
-  public static async updateBotProxy(botId, group, endpoint) {
+  public async updateBotProxy(botId, group, endpoint) {
     const baseUrl = `https://management.azure.com/`;
     const username = GBConfigService.get('CLOUD_USERNAME');
     const password = GBConfigService.get('CLOUD_PASSWORD');
     const subscriptionId = GBConfigService.get('CLOUD_SUBSCRIPTIONID');
 
-    const accessToken = await GBAdminService.getADALTokenFromUsername(
-      username,
-      password
-    );
+    const accessToken = await GBAdminService.getADALTokenFromUsername(username, password);
     const httpClient = new ServiceClient();
 
     const parameters = {
@@ -207,22 +218,18 @@ export class AzureDeployerService extends GBService {
 
     const query = `subscriptions/${subscriptionId}/resourceGroups/${group}/providers/${
       this.provider
-    }/botServices/${botId}?api-version=${AzureDeployerService.apiVersion}`;
-    const url = UrlJoin(baseUrl, query);
-    const req = this.createRequestObject(
-      url,
-      accessToken,
-      'PATCH',
-      JSON.stringify(parameters)
-    );
+      }/botServices/${botId}?api-version=${this.apiVersion}`;
+    const url = urlJoin(baseUrl, query);
+    const req = AzureDeployerService.createRequestObject(url, accessToken, 'PATCH', JSON.stringify(parameters));
     const res = await httpClient.sendRequest(req);
-    if (!(res.bodyAsJson as any).id) {
+    // CHECK
+    if (!JSON.parse(res.bodyAsText).id) {
       throw res.bodyAsText;
     }
-    logger.info(`Bot proxy updated at: ${endpoint}.`);
+    GBLog.info(`Bot proxy updated at: ${endpoint}.`);
   }
 
-  public static async openStorageFirewall(groupName, serverName) {
+  public async openStorageFirewall(groupName, serverName) {
     const username = GBConfigService.get('CLOUD_USERNAME');
     const password = GBConfigService.get('CLOUD_PASSWORD');
     const subscriptionId = GBConfigService.get('CLOUD_SUBSCRIPTIONID');
@@ -239,19 +246,6 @@ export class AzureDeployerService extends GBService {
     await storageClient.firewallRules.createOrUpdate(groupName, serverName, 'gb', params);
   }
 
-  public static createRequestObject(url: string, accessToken: string, verb: HttpMethods, body: string) {
-    const req = new WebResource();
-    req.method = verb;
-    req.url = url;
-    req.headers = {};
-    req.headers['Content-Type'] = 'application/json';
-    req.headers['accept-language'] = '*';
-    req.headers.Authorization = 'Bearer ' + accessToken;
-    req.body = body;
-
-    return req;
-  }
-
   public async deployFarm(
     proxyAddress: string,
     instance: IGBInstance,
@@ -261,22 +255,20 @@ export class AzureDeployerService extends GBService {
     const culture = 'en-us';
 
     this.initServices(credentials, subscriptionId);
-
     const spinner = new Spinner('%s');
     spinner.start();
     spinner.setSpinnerString('|/-\\');
-
     let keys: any;
     const name = instance.botId;
 
-    logger.info(`Deploying Deploy Group (It may take a few minutes)...`);
+    GBLog.info(`Deploying Deploy Group (It may take a few minutes)...`);
     await this.createDeployGroup(name, instance.cloudLocation);
 
-    logger.info(`Deploying Bot Server...`);
+    GBLog.info(`Deploying Bot Server...`);
     const serverFarm = await this.createHostingPlan(name, `${name}-server-plan`, instance.cloudLocation);
     await this.createServer(serverFarm.id, name, `${name}-server`, instance.cloudLocation);
 
-    logger.info(`Deploying Bot Storage...`);
+    GBLog.info(`Deploying Bot Storage...`);
     const administratorLogin = `sa${GBAdminService.getRndReadableIdentifier()}`;
     const administratorPassword = GBAdminService.getRndPassword();
     const storageServer = `${name.toLowerCase()}-storage-server`;
@@ -296,7 +288,7 @@ export class AzureDeployerService extends GBService {
     instance.storageDialect = 'mssql';
     instance.storageServer = storageServer;
 
-    logger.info(`Deploying Search...`);
+    GBLog.info(`Deploying Search...`);
     const searchName = `${name}-search`.toLowerCase();
     await this.createSearch(name, searchName, instance.cloudLocation);
     const searchKeys = await this.searchClient.adminKeys.get(name, searchName);
@@ -304,27 +296,28 @@ export class AzureDeployerService extends GBService {
     instance.searchIndex = 'azuresql-index';
     instance.searchIndexer = 'azuresql-indexer';
     instance.searchKey = searchKeys.primaryKey;
-    this.deployer.rebuildIndex(instance);
+    this.deployer.rebuildIndex(instance, this.deployer);
 
-    logger.info(`Deploying Speech...`);
+    GBLog.info(`Deploying Speech...`);
     const speech = await this.createSpeech(name, `${name}-speech`, instance.cloudLocation);
     keys = await this.cognitiveClient.accounts.listKeys(name, speech.name);
-    instance.speechKeyEndpoint = speech.endpoint;
+    instance.speechEndpoint = speech.endpoint;
     instance.speechKey = keys.key1;
 
-    logger.info(`Deploying SpellChecker...`);
-    const spellChecker = await this.createSpellChecker(name, `${name}-spellchecker`, instance.cloudLocation);
+    GBLog.info(`Deploying SpellChecker...`);
+    const spellChecker = await this.createSpellChecker(name, `${name}-spellchecker`);
     keys = await this.cognitiveClient.accounts.listKeys(name, spellChecker.name);
-    instance.spellCheckerKey = keys.key1;
-    instance.spellCheckerEndpoint = spellChecker.endpoint;
+    instance.spellcheckerKey = keys.key1;
+    instance.spellcheckerEndpoint = spellChecker.endpoint;
 
-    logger.info(`Deploying Text Analytics...`);
+    GBLog.info(`Deploying Text Analytics...`);
     const textAnalytics = await this.createTextAnalytics(name, `${name}-textanalytics`, instance.cloudLocation);
     keys = await this.cognitiveClient.accounts.listKeys(name, textAnalytics.name);
-    instance.textAnalyticsEndpoint = textAnalytics.endpoint;
+
+    instance.textAnalyticsEndpoint = textAnalytics.endpoint.replace(`/text/analytics/v2.0`, '');
     instance.textAnalyticsKey = keys.key1;
 
-    logger.info(`Deploying NLP...`);
+    GBLog.info(`Deploying NLP...`);
     const nlp = await this.createNLP(name, `${name}-nlp`, instance.cloudLocation);
     keys = await this.cognitiveClient.accounts.listKeys(name, nlp.name);
     const nlpAppId = await this.createNLPService(name, name, instance.cloudLocation, culture, instance.nlpAuthoringKey);
@@ -333,8 +326,8 @@ export class AzureDeployerService extends GBService {
     instance.nlpKey = keys.key1;
     instance.nlpAppId = nlpAppId;
 
-    logger.info(`Deploying Bot...`);
-    instance.botEndpoint = AzureDeployerService.defaultEndPoint;
+    GBLog.info(`Deploying Bot...`);
+    instance.botEndpoint = this.defaultEndPoint;
 
     instance = await this.internalDeployBot(
       instance,
@@ -347,27 +340,27 @@ export class AzureDeployerService extends GBService {
       'global',
       instance.nlpAppId,
       instance.nlpKey,
-      instance.appId,
-      instance.appPassword,
+      instance.marketplaceId,
+      instance.marketplacePassword,
       instance.cloudSubscriptionId
     );
 
     spinner.stop();
+
     return instance;
   }
 
   public async deployToCloud(
-    title,
-    username,
-    password,
-    cloudLocation,
-    authoringKey,
-    appId,
-    appPassword,
-    subscriptionId
+    title: string,
+    username: string,
+    password: string,
+    cloudLocation: string,
+    authoringKey: string,
+    appId: string,
+    appPassword: string,
+    subscriptionId: string
   ) {
-
-    const instance: IGBInstance = {};
+    const instance = <IGBInstance>{};
 
     instance.botId = title;
     instance.cloudUsername = username;
@@ -380,10 +373,9 @@ export class AzureDeployerService extends GBService {
     instance.adminPass = GBAdminService.getRndPassword();
 
     const credentials = await GBAdminService.getADALCredentialsFromUsername(username, password);
-    this.deployFarm(`http://${instance.botId}.azurewebsites.net`, instance, credentials, subscriptionId);
-
-    // TODO: Copy github to webapp.
-    //const status = await git.status();
+    // tslint:disable-next-line:no-http-string
+    const url = `http://${instance.botId}.azurewebsites.net`;
+    this.deployFarm(url, instance, credentials, subscriptionId);
   }
 
   private initServices(credentials: any, subscriptionId: string) {
@@ -393,19 +385,6 @@ export class AzureDeployerService extends GBService {
     this.cognitiveClient = new CognitiveServicesManagementClient(credentials, subscriptionId);
     this.searchClient = new SearchManagementClient(credentials, subscriptionId);
     this.accessToken = credentials.tokenCache._entries[0].accessToken;
-  }
-
-  private async updateWebisteConfig(group, serverFarmId, name, location) {
-    const siteConfig = {
-      location: location,
-      serverFarmId: serverFarmId,
-      numberOfWorkers: 1,
-      phpVersion: '5.5'
-    };
-
-    // TODO: Copy .env to app settings.
-
-    return this.webSiteClient.webApps.createOrUpdateConfiguration(group, name, siteConfig);
   }
 
   private async createStorageServer(group, name, administratorLogin, administratorPassword, serverName, location) {
@@ -420,22 +399,16 @@ export class AzureDeployerService extends GBService {
   }
 
   private async registerProviders(subscriptionId, baseUrl, accessToken) {
-    const query = `subscriptions/${subscriptionId}/providers/${
-      AzureDeployerService.provider
-    }/register?api-version=2018-02-01`;
-    const requestUrl = UrlJoin(baseUrl, query);
+    const query = `subscriptions/${subscriptionId}/providers/${this.provider}/register?api-version=2018-02-01`;
+    const requestUrl = urlJoin(baseUrl, query);
 
     const req = new WebResource();
     req.method = 'POST';
     req.url = requestUrl;
-    req.headers = {};
+    req.headers = <any>{};
     req.headers['Content-Type'] = 'application/json; charset=utf-8';
     req.headers['accept-language'] = '*';
-    req.headers.Authorization = 'Bearer ' + accessToken;
-
-    const httpClient = new ServiceClient();
-    const res = await httpClient.sendRequest(req);
-    // TODO: Check res for error.
+    (req.headers as any).Authorization = `Bearer ${accessToken}`;
   }
 
   /**
@@ -455,7 +428,7 @@ export class AzureDeployerService extends GBService {
     appId,
     appPassword,
     subscriptionId
-  ) {
+  ): Promise<IGBInstance> {
     return new Promise(async (resolve, reject) => {
       const baseUrl = `https://management.azure.com/`;
       await this.registerProviders(subscriptionId, baseUrl, accessToken);
@@ -487,25 +460,27 @@ export class AzureDeployerService extends GBService {
 
       const httpClient = new ServiceClient();
       let query = `subscriptions/${subscriptionId}/resourceGroups/${group}/providers/${
-        AzureDeployerService.provider
-      }/botServices/${botId}?api-version=${AzureDeployerService.apiVersion}`;
-      let url = UrlJoin(baseUrl, query);
+        this.provider
+        }/botServices/${botId}?api-version=${this.apiVersion}`;
+      let url = urlJoin(baseUrl, query);
       let req = AzureDeployerService.createRequestObject(url, accessToken, 'PUT', JSON.stringify(parameters));
       const res = await httpClient.sendRequest(req);
-      if (!(res.bodyAsJson as any).id) {
+      if (!JSON.parse(res.bodyAsText).id) {
         reject(res.bodyAsText);
+
         return;
       }
 
       setTimeout(async () => {
         try {
+          //tslint:disable-next-line:max-line-length
           query = `subscriptions/${subscriptionId}/resourceGroups/${group}/providers/Microsoft.BotService/botServices/${botId}/channels/WebChatChannel/listChannelWithKeys?api-version=${
-            AzureDeployerService.apiVersion
-          }`;
-          url = UrlJoin(baseUrl, query);
-          req = AzureDeployerService.createRequestObject(url, accessToken, 'GET', JSON.stringify(parameters));
+            this.apiVersion
+            }`;
+          url = urlJoin(baseUrl, query);
+          req = AzureDeployerService.createRequestObject(url, accessToken, 'POST', JSON.stringify(parameters));
           const resChannel = await httpClient.sendRequest(req);
-          const key = (resChannel.bodyAsJson as any).properties.properties.sites[0].key;
+          const key = JSON.parse(resChannel.bodyAsText).properties.properties.sites[0].key;
           instance.webchatKey = key;
           resolve(instance);
         } catch (error) {
@@ -529,8 +504,8 @@ export class AzureDeployerService extends GBService {
     };
 
     const body = JSON.stringify(parameters);
-    const apps = await this.makeNlpRequest(location, authoringKey, null, 'GET', 'apps');
-    const app = (apps.bodyAsJson as any).filter(x => x.name == name)[0];
+    const apps = await this.makeNlpRequest(location, authoringKey, undefined, 'GET', 'apps');
+    const app = JSON.parse(apps.bodyAsText).filter(x => x.name === name)[0];
     let id: string;
     if (!app) {
       const res = await this.makeNlpRequest(location, authoringKey, body, 'POST', 'apps');
@@ -552,10 +527,9 @@ export class AzureDeployerService extends GBService {
     const req = new WebResource();
     req.method = method;
     req.url = `https://${location}.api.cognitive.microsoft.com/luis/api/v2.0/${resource}`;
-    req.headers = {};
-    req.headers['Content-Type'] = 'application/json';
-    req.headers['accept-language'] = '*';
-    req.headers['Ocp-Apim-Subscription-Key'] = authoringKey;
+    req.headers.set('Content-Type', 'application/json');
+    req.headers.set('accept-language', '*');
+    req.headers.set('Ocp-Apim-Subscription-Key', authoringKey);
     req.body = body;
     const httpClient = new ServiceClient();
 
@@ -601,7 +575,7 @@ export class AzureDeployerService extends GBService {
     return await this.createCognitiveServices(group, name, location, 'LUIS');
   }
 
-  private async createSpellChecker(group, name, location): Promise<CognitiveServicesAccount> {
+  private async createSpellChecker(group, name): Promise<CognitiveServicesAccount> {
     return await this.createCognitiveServices(group, name, 'global', 'Bing.SpellCheck.v7');
   }
 
@@ -611,6 +585,7 @@ export class AzureDeployerService extends GBService {
 
   private async createDeployGroup(name, location) {
     const params = { location: location };
+
     return this.resourceClient.resourceGroups.createOrUpdate(name, params);
   }
 
@@ -633,7 +608,7 @@ export class AzureDeployerService extends GBService {
       location: location,
       serverFarmId: farmId
     };
+
     return this.webSiteClient.webApps.createOrUpdate(group, name, parameters);
   }
-  
 }
