@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -36,30 +36,33 @@
 
 'use strict';
 
-const UrlJoin = require('url-join');
-
-import { GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
-
+import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
 import { Sequelize } from 'sequelize-typescript';
 import { ConsoleDirectLine } from './services/ConsoleDirectLine';
 
+/**
+ * Package for console.glib.
+ */
 export class GBConsolePackage implements IGBPackage {
-  public sysPackages: IGBPackage[] = null;
+  public sysPackages: IGBPackage[];
   public channel: ConsoleDirectLine;
-
-  public loadPackage(core: IGBCoreService, sequelize: Sequelize): void {
+  public getDialogs(min: GBMinInstance) {
+    GBLog.verbose(`getDialogs called.`);
   }
-
+  public loadPackage(core: IGBCoreService, sequelize: Sequelize): void {
+    GBLog.verbose(`loadPackage called.`);
+  }
   public unloadPackage(core: IGBCoreService): void {
+    GBLog.verbose(`unloadPackage called.`);
+  }
+  public unloadBot(min: GBMinInstance): void {
+    GBLog.verbose(`unloadBot called.`);
+  }
+  public onNewSession(min: GBMinInstance, step: GBDialogStep): void {
+    GBLog.verbose(`onNewSession called.`);
   }
 
   public loadBot(min: GBMinInstance): void {
     this.channel = new ConsoleDirectLine(min.instance.webchatKey);
-  }
-
-  public unloadBot(min: GBMinInstance): void {
-  }
-
-  public onNewSession(min: GBMinInstance, step: any): void {
   }
 }
