@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -36,13 +36,11 @@
 
 'use strict';
 
-const logger = require('../../../src/logger');
-import { MessageFactory } from 'botbuilder';
+import { MessageFactory, RecognizerResult } from 'botbuilder';
 import { LuisRecognizer } from 'botbuilder-ai';
-import { GBMinInstance, IGBConversationalService } from 'botlib';
+import { GBDialogStep, GBLog, GBMinInstance, IGBConversationalService, IGBCoreService } from 'botlib';
 import { AzureText } from 'pragmatismo-io-framework';
 import { Messages } from '../strings';
-import { GBCoreService } from './GBCoreService';
 const Nexmo = require('nexmo');
 
 export interface LanguagePickerSettings {
@@ -50,18 +48,22 @@ export interface LanguagePickerSettings {
   supportedLocales?: string[];
 }
 
+/**
+ * Provides basic services for handling messages and dispatching to back-end
+ * services like NLP or Search.
+ */
 export class GBConversationalService implements IGBConversationalService {
-  public coreService: GBCoreService;
+  public coreService: IGBCoreService;
 
-  constructor(coreService: GBCoreService) {
+  constructor(coreService: IGBCoreService) {
     this.coreService = coreService;
   }
 
-  public getCurrentLanguage(step: any) {
+  public getCurrentLanguage(step: GBDialogStep) {
     return step.context.activity.locale;
   }
 
-  public async sendEvent(step: any, name: string, value: any): Promise<any> {
+  public async sendEvent(step: GBDialogStep, name: string, value: Object): Promise<any> {
     if (step.context.activity.channelId === 'webchat') {
       const msg = MessageFactory.text('');
       msg.value = value;
@@ -72,6 +74,7 @@ export class GBConversationalService implements IGBConversationalService {
     }
   }
 
+  // tslint:disable:no-unsafe-any due to Nexmo.
   public async sendSms(min: GBMinInstance, mobile: string, text: string): Promise<any> {
     return new Promise(
       (resolve: any, reject: any): any => {
@@ -79,6 +82,7 @@ export class GBConversationalService implements IGBConversationalService {
           apiKey: min.instance.smsKey,
           apiSecret: min.instance.smsSecret
         });
+        // tslint:disable-next-line:no-unsafe-any
         nexmo.message.sendSms(min.instance.smsServiceNumber, mobile, text, (err, data) => {
           if (err) {
             reject(err);
@@ -89,11 +93,12 @@ export class GBConversationalService implements IGBConversationalService {
       }
     );
   }
+  // tslint:enable:no-unsafe-any
 
-  public async routeNLP(step: any, min: GBMinInstance, text: string): Promise<boolean> {
+  public async routeNLP(step: GBDialogStep, min: GBMinInstance, text: string): Promise<boolean> {
     // Invokes LUIS.
 
-    let endpoint = min.instance.nlpEndpoint.replace('/luis/v2.0', '');
+    const endpoint = min.instance.nlpEndpoint.replace('/luis/v2.0', '');
 
     const model = new LuisRecognizer({
       applicationId: min.instance.nlpAppId,
@@ -101,38 +106,42 @@ export class GBConversationalService implements IGBConversationalService {
       endpoint: endpoint
     });
 
-    let nlp: any;
+    let nlp: RecognizerResult;
     try {
       nlp = await model.recognize(step.context);
     } catch (error) {
+      // tslint:disable:no-unsafe-any
       if (error.statusCode === 404) {
-        logger.warn('NLP application still not publish and there are no other options for answering.');
+        GBLog.warn('NLP application still not publish and there are no other options for answering.');
 
         return Promise.resolve(false);
       } else {
         const msg = `Error calling NLP, check if you have a published model and assigned keys. Error: ${
           error.statusCode ? error.statusCode : ''
-        } ${error.message}`;
+        } {error.message; }`;
 
         return Promise.reject(new Error(msg));
       }
+      // tslint:enable:no-unsafe-any
     }
 
     // Resolves intents returned from LUIS.
 
     const topIntent = LuisRecognizer.topIntent(nlp);
-    if (topIntent) {
+    if (topIntent !== undefined) {
       const intent = topIntent;
-      const entity = nlp.entities && nlp.entities.length > 0 ? nlp.entities[0].entity.toUpperCase() : null;
+      // tslint:disable:no-unsafe-any
+      const firstEntity = nlp.entities && nlp.entities.length > 0 ? nlp.entities[0].entity.toUpperCase() : undefined;
+      // tslint:ensable:no-unsafe-any
 
       if (intent === 'None') {
         return Promise.resolve(false);
       }
 
-      logger.info(`NLP called: ${intent}, ${entity}`);
+      GBLog.info(`NLP called: ${intent} ${firstEntity}`);
 
       try {
-        await step.replaceDialog(`/${intent}`, nlp.entities);
+        await step.replaceDialog(` /${intent}`, nlp.entities);
 
         return Promise.resolve(true);
       } catch (error) {
@@ -145,7 +154,7 @@ export class GBConversationalService implements IGBConversationalService {
     return Promise.resolve(false);
   }
 
-  public async checkLanguage(step, min, text) {
+  public async checkLanguage(step: GBDialogStep, min, text) {
     const locale = await AzureText.getLocale(min.instance.textAnalyticsKey, min.instance.textAnalyticsEndpoint, text);
     if (locale !== step.context.activity.locale.split('-')[0]) {
       switch (locale) {
@@ -158,7 +167,7 @@ export class GBConversationalService implements IGBConversationalService {
           await step.context.sendActivity(Messages[locale].changing_language);
           break;
         default:
-          await step.context.sendActivity(`Unknown language: ${locale}`);
+          await step.context.sendActivity(`; Unknown; language: $;{locale;}`);
           break;
       }
     }
