@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -36,32 +36,37 @@
 
 'use strict';
 
-const UrlJoin = require('url-join');
-import { GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
+import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
+import urlJoin = require('url-join');
 import { FeedbackDialog } from './dialogs/FeedbackDialog';
 import { QualityDialog } from './dialogs/QualityDialog';
 import { GuaribasQuestionAlternate } from './models/index';
 
 import { Sequelize } from 'sequelize-typescript';
 
+/**
+ * Package for customer-satisfaction.gblib.
+ */
 export class GBCustomerSatisfactionPackage implements IGBPackage {
-  public sysPackages: IGBPackage[] = null;
-  public loadPackage(core: IGBCoreService, sequelize: Sequelize): void {
-    core.sequelize.addModels([
-      GuaribasQuestionAlternate
-    ]);
+  public sysPackages: IGBPackage[];
+  public getDialogs(min: GBMinInstance) {
+    GBLog.verbose(`getDialogs called.`);
   }
   public unloadPackage(core: IGBCoreService): void {
+    GBLog.verbose(`unloadPackage called.`);
+  }
+  public unloadBot(min: GBMinInstance): void {
+    GBLog.verbose(`unloadBot called.`);
+  }
+  public onNewSession(min: GBMinInstance, step: GBDialogStep): void {
+    GBLog.verbose(`onNewSession called.`);
+  }
 
+  public loadPackage(core: IGBCoreService, sequelize: Sequelize): void {
+    core.sequelize.addModels([GuaribasQuestionAlternate]);
   }
   public loadBot(min: GBMinInstance): void {
     FeedbackDialog.setup(min.bot, min);
     QualityDialog.setup(min.bot, min);
-  }
-  public unloadBot(min: GBMinInstance): void {
-
-  }
-  public onNewSession(min: GBMinInstance, step: any): void {
-
   }
 }
