@@ -1,266 +1,218 @@
-/*****************************************************************************\
-|                                               ( )_  _                       |
-|    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
-|   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
-|   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
-|   | |                ( )_) |                                                |
-|   (_)                 \___/'                                                |
-|                                                                             |
-| General Bots Copyright (c) Pragmatismo.io. All rights reserved.             |
-| Licensed under the AGPL-3.0.                                                |
-|                                                                             |
-| According to our dual licensing model, this program can be used either      |
-| under the terms of the GNU Affero General Public License, version 3,        |
-| or under a proprietary license.                                             |
-|                                                                             |
-| The texts of the GNU Affero General Public License with an additional       |
-| permission and of our proprietary license can be found at and               |
-| in the LICENSE file you have received along with this program.              |
-|                                                                             |
-| This program is distributed in the hope that it will be useful,             |
-| but WITHOUT ANY WARRANTY, without even the implied warranty of              |
-| MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                |
-| GNU Affero General Public License for more details.                         |
-|                                                                             |
-| "General Bots" is a registered trademark of Pragmatismo.io.                 |
-| The licensing of the program under the AGPLv3 does not imply a              |
-| trademark license. Therefore any rights, title and interest in              |
-| our trademarks remain entirely with us.                                     |
-|                                                                             |
-\*****************************************************************************/
+import urlJoin = require('url-join');
 
-const Path = require('path');
-const Fs = require('fs');
-const _ = require('lodash');
-const Parse = require('csv-parse');
-const Async = require('async');
-const UrlJoin = require('url-join');
-const logger = require('../../../src/logger');
 const Swagger = require('swagger-client');
 const rp = require('request-promise');
+import { GBLog, GBService } from 'botlib';
 import * as request from 'request-promise-native';
 
-import { GBService, GBServiceCallback, IGBInstance } from 'botlib';
-
+/**
+ * Support for Whatsapp.
+ */
 export class WhatsappDirectLine extends GBService {
+  public pollInterval = 1000;
+  public directLineClientName = 'DirectLineClient';
+  public directLineSpecUrl = 'https://docs.botframework.com/en-us/restapi/directline3/swagger.json';
 
-    public pollInterval = 1000;
-    public directLineClientName = 'DirectLineClient';
-    public directLineSpecUrl = 'https://docs.botframework.com/en-us/restapi/directline3/swagger.json';
+  public directLineClient: any;
+  public whatsappServiceKey: string;
+  public whatsappServiceNumber: string;
+  public whatsappServiceUrl: string;
+  public whatsappServiceWebhookUrl: string;
+  public botId: string;
+  public watermark: string;
 
-    public directLineClient: any;
-    public whatsappServiceKey: string;
-    public whatsappServiceNumber: string;
-    public whatsappServiceUrl: string;
-    public whatsappServiceWebhookUrl: string;
-    public botId: string;
-    public watermark: string = null;
+  public conversationIds = {};
 
-    public conversationIds = {};
+  constructor(
+    botId,
+    directLineSecret,
+    whatsappServiceKey,
+    whatsappServiceNumber,
+    whatsappServiceUrl,
+    whatsappServiceWebhookUrl
+  ) {
+    super();
 
-    constructor(botId, directLineSecret, whatsappServiceKey, whatsappServiceNumber, whatsappServiceUrl, whatsappServiceWebhookUrl) {
+    this.botId = botId;
+    this.whatsappServiceKey = whatsappServiceKey;
+    this.whatsappServiceNumber = whatsappServiceNumber;
+    this.whatsappServiceUrl = whatsappServiceUrl;
+    this.whatsappServiceWebhookUrl = whatsappServiceWebhookUrl;
 
-        super();
-
-        this.botId = botId;
-        this.whatsappServiceKey = whatsappServiceKey;
-        this.whatsappServiceNumber = whatsappServiceNumber;
-        this.whatsappServiceUrl = whatsappServiceUrl;
-        this.whatsappServiceWebhookUrl = whatsappServiceWebhookUrl;
-
-        // TODO: Migrate to Swagger 3.
-        this.directLineClient = rp(this.directLineSpecUrl)
-            .then((spec) => {
-                return new Swagger({
-                    spec: JSON.parse(spec.trim()),
-                    usePromise: true
-                });
-            })
-            .then(async (client) => {
-                client.clientAuthorizations.add('AuthorizationBotConnector',
-                                                new Swagger.ApiKeyAuthorization('Authorization', 'Bearer ' +
-                        directLineSecret,                                       'header'));
-
-                const options = {
-                    method: 'POST',
-                    url: UrlJoin(this.whatsappServiceUrl, 'webhook'),
-                    qs:
-                        {
-                            token: this.whatsappServiceKey,
-                            webhookUrl: `${this.whatsappServiceWebhookUrl}/instances/${this.botId}/whatsapp`,
-                            set: true
-                        },
-                    headers:
-                        {
-                            'cache-control': 'no-cache'
-                        }
-                };
-
-                try {
-                    const result = await request.post(options);
-                    logger.info(result);
-                } catch (error) {
-                    logger.error('Error initializing 3rd party Whatsapp provider.', error);
-                }
-
-                return client;
-            })
-            .catch((err) => {
-                logger.error('Error initializing DirectLine client', err);
-            });
-
-    }
-
-    public received(req, res) {
-        const text = req.body.messages[0].body;
-        const from = req.body.messages[0].author.split('@')[0];
-        const fromName = req.body.messages[0].senderName;
-
-        if (req.body.messages[0].fromMe) {
-            return; // Exit here.
-        }
-
-        logger.info(`GBWhatsapp: Hook called. from: ${from}(${fromName}), text: ${text})`);
-
-        const conversationId = this.conversationIds[from];
-
-        this.directLineClient.then((client) => {
-
-            if (this.conversationIds[from] == null) {
-
-                logger.info(`GBWhatsapp: Starting new conversation on Bot.`);
-                client.Conversations.Conversations_StartConversation()
-                    .then((response) => {
-                        return response.obj.conversationId;
-                    })
-                    .then((conversationId) => {
-
-                        this.conversationIds[from] = conversationId;
-                        this.inputMessage(client, conversationId, text,
-                                          from, fromName);
-
-                        this.pollMessages(client, conversationId, from, fromName);
-                    })
-                    .catch((err) => {
-                        console.error('Error starting conversation', err);
-                    });
-
-            } else {
-                this.inputMessage(client, conversationId, text,
-                                  from, fromName);
-            }
-            res.end();
+    this.directLineClient = rp(this.directLineSpecUrl)
+      .then(spec => {
+        return new Swagger({
+          spec: JSON.parse(spec.trim()),
+          usePromise: true
         });
-    }
+      })
+      .then(async client => {
+        client.clientAuthorizations.add(
+          'AuthorizationBotConnector',
+          new Swagger.ApiKeyAuthorization('Authorization', `Bearer ${directLineSecret}`, 'header')
+        );
 
-    public inputMessage(client, conversationId, text, from, fromName) {
-
-        client.Conversations.Conversations_PostActivity(
-            {
-                conversationId: conversationId,
-                activity: {
-                    textFormat: 'plain',
-                    text: text,
-                    type: 'message',
-                    from: {
-                        id: from,
-                        name: fromName
-                    },
-                    replyToId: from
-                }
-            }).catch((err) => {
-                logger.error(`GBWhatsapp: Error receiving message: ${err}.`);
-            });
-
-    }
-
-    public pollMessages(client, conversationId, from, fromName) {
-
-        logger.info(`GBWhatsapp: Starting polling message for conversationId:
-        ${conversationId}.`);
-
-        setInterval(() => {
-            client.Conversations.Conversations_GetActivities({
-                conversationId:
-                    conversationId, watermark: this.watermark
-            })
-                .then((response) => {
-                    this.watermark = response.obj.watermark;
-                    return response.obj.activities;
-                })
-                .then((activities) => {
-                    this.printMessages(activities, conversationId, from, fromName);
-                });
-        },          this.pollInterval);
-    }
-
-    public printMessages(activities, conversationId, from, fromName) {
-
-        if (activities && activities.length) {
-
-            // Ignore own messages.
-
-		activities = activities.filter((m) => (m.from.id === 'GeneralBots')  && m.type === 'message');
-
-  if (activities.length) {
-
-                // Print other messages.
-
-                activities.forEach(activity => {
-                    this.printMessage(activity, conversationId, from, fromName);
-                });
-            }
-        }
-    }
-
-    public printMessage(activity, conversationId, from, fromName) {
-
-        let output = '';
-
-        if (activity.text) {
-            logger.info(`GBWhatsapp: MSG: ${activity.text}`);
-            output = activity.text;
-        }
-
-        if (activity.attachments) {
-            activity.attachments.forEach((attachment) => {
-                switch (attachment.contentType) {
-                    case 'application/vnd.microsoft.card.hero':
-                        output += `\n${this.renderHeroCard(attachment)}`;
-                        break;
-
-                    case 'image/png':
-                        logger.info('Opening the requested image ' + attachment.contentUrl);
-                        output += `\n${attachment.contentUrl}`;
-                        break;
-                }
-            });
-        }
-
-        this.sendToDevice(conversationId, from, fromName, output);
-    }
-
-    public renderHeroCard(attachment) {
-        return `${attachment.content.title} - ${attachment.content.text}`;
-    }
-
-    public async sendToDevice(conversationId, to, toName, msg) {
         const options = {
-            method: 'POST',
-            url: UrlJoin(this.whatsappServiceUrl, 'message'),
-            qs:
-                {
-                    token: this.whatsappServiceKey,
-                    phone: to,
-                    body: msg
-                },
-            headers:
-                {
-                    'cache-control': 'no-cache'
-                }
+          method: 'POST',
+          url: urlJoin(this.whatsappServiceUrl, 'webhook'),
+          qs: {
+            token: this.whatsappServiceKey,
+            webhookUrl: `${this.whatsappServiceWebhookUrl}/instances/${this.botId}/whatsapp`,
+            set: true
+          },
+          headers: {
+            'cache-control': 'no-cache'
+          }
         };
 
-        const result = await request.get(options);
+        try {
+          const result = request.post(options);
+          GBLog.info(result);
+        } catch (error) {
+          GBLog.error(`Error initializing 3rd party Whatsapp provider(1) ${error}`);
+        }
+
+        return client;
+      })
+      .catch(err => {
+        GBLog.error(`Error initializing 3rd party Whatsapp provider(2) ${err}`);
+      });
+  }
+
+  public received(req, res) {
+    const text = req.body.messages[0].body;
+    const from = req.body.messages[0].author.split('@')[0];
+    const fromName = req.body.messages[0].senderName;
+
+    if (req.body.messages[0].fromMe) {
+      return; // Exit here.
     }
+
+    GBLog.info(`GBWhatsapp: Hook called. from: ${from}(${fromName}), text: ${text})`);
+
+    const conversationId = this.conversationIds[from];
+
+    this.directLineClient.then(client => {
+      if (this.conversationIds[from] === undefined) {
+        GBLog.info(`GBWhatsapp: Starting new conversation on Bot.`);
+        client.Conversations.Conversations_StartConversation()
+          .then(response => {
+            return response.obj.conversationId;
+          })
+          .then(generatedConversationId => {
+            this.conversationIds[from] = generatedConversationId;
+            this.inputMessage(client, generatedConversationId, text, from, fromName);
+
+            this.pollMessages(client, generatedConversationId, from, fromName);
+          })
+          .catch(err => {
+            GBLog.error(`Error starting conversation ${err}`);
+          });
+      } else {
+        this.inputMessage(client, conversationId, text, from, fromName);
+      }
+      res.end();
+    });
+  }
+
+  public inputMessage(client, conversationId, text, from, fromName) {
+    client.Conversations.Conversations_PostActivity({
+      conversationId: conversationId,
+      activity: {
+        textFormat: 'plain',
+        text: text,
+        type: 'message',
+        from: {
+          id: from,
+          name: fromName
+        },
+        replyToId: from
+      }
+    }).catch(err => {
+      GBLog.error(`GBWhatsapp: Error receiving message: ${err}.`);
+    });
+  }
+
+  public pollMessages(client, conversationId, from, fromName) {
+    GBLog.info(`GBWhatsapp: Starting polling message for conversationId:
+        ${conversationId}.`);
+
+    setInterval(() => {
+      client.Conversations.Conversations_GetActivities({
+        conversationId: conversationId,
+        watermark: this.watermark
+      })
+        .then(response => {
+          this.watermark = response.obj.watermark;
+
+          return response.obj.activities;
+        })
+        .then(activities => {
+          this.printMessages(activities, conversationId, from, fromName);
+        });
+    },          this.pollInterval);
+  }
+
+  public printMessages(activities, conversationId, from, fromName) {
+    if (activities && activities.length) {
+      // Ignore own messages.
+
+      activities = activities.filter(m => m.from.id === 'GeneralBots' && m.type === 'message');
+
+      if (activities.length) {
+        // Print other messages.
+
+        activities.forEach(activity => {
+          this.printMessage(activity, conversationId, from, fromName);
+        });
+      }
+    }
+  }
+
+  public printMessage(activity, conversationId, from, fromName) {
+    let output = '';
+
+    if (activity.text) {
+      GBLog.info(`GBWhatsapp: MSG: ${activity.text}`);
+      output = activity.text;
+    }
+
+    if (activity.attachments) {
+      activity.attachments.forEach(attachment => {
+        switch (attachment.contentType) {
+          case 'application/vnd.microsoft.card.hero':
+            output += `\n${this.renderHeroCard(attachment)}`;
+            break;
+
+          case 'image/png':
+            GBLog.info(`Opening the requested image ${attachment.contentUrl}`);
+            output += `\n${attachment.contentUrl}`;
+            break;
+          default:
+            GBLog.info(`Unknown content type: ${attachment.contentType}`);
+        }
+      });
+    }
+
+    this.sendToDevice(from, output);
+  }
+
+  public renderHeroCard(attachment) {
+    return `${attachment.content.title} - ${attachment.content.text}`;
+  }
+
+  public async sendToDevice(to, msg) {
+    const options = {
+      method: 'POST',
+      url: urlJoin(this.whatsappServiceUrl, 'message'),
+      qs: {
+        token: this.whatsappServiceKey,
+        phone: to,
+        body: msg
+      },
+      headers: {
+        'cache-control': 'no-cache'
+      }
+    };
+  }
 }
